requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
pandas>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
hypothesis>=6.80.0